:wrench: Move icons in the same package fyndiq-icons
