+++ packages/fyndiq-component-dropdown/package.json
@@ -11,7 +11,7 @@
   "homepage": "http://developers.fyndiq.com/fyndiq-ui/?selectedKind=Dropdown&selectedStory=default",
   "dependencies": {
     "fyndiq-component-button": "^2.0.1",
-    "fyndiq-icon-arrow": "^2.0.1",
+    "fyndiq-icons": "*",
     "fyndiq-styles-colors": "^2.0.0"
   },
   "peerDependencies": {
