+++ packages/fyndiq-component-dropdown/src/index.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import Button from 'fyndiq-component-button'
-import Arrow from 'fyndiq-icon-arrow'
+import { Arrow } from 'fyndiq-icons'
+
 import styles from '../styles.css'
 
 class Dropdown extends React.Component {
