--- packages/fyndiq-icon-checkmark/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "fyndiq-icon-checkmark",
-  "version": "2.0.1",
-  "author": "Thibaut REMY <thibaut.remy@fyndiq.com>",
-  "main": "lib/index.js",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/fyndiq/fyndiq-ui.git"
-  },
-  "homepage": "https://github.com/fyndiq/fyndiq-ui",
-  "dependencies": {
-    "fyndiq-styles-colors": "^2.0.0"
-  },
-  "peerDependencies": {
-    "prop-types": "^15.5.10",
-    "react": "^15.6.1 || ^16.0.0"
-  }
-}
