--- packages/fyndiq-icon-checkmark/src/index.test.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import { shallow } from 'enzyme'
-
-import Checkmark from './'
-
-describe('fyndiq-icon-brand', () => {
-  it('should be rendered without props', () => {
-    expect(shallow(<Checkmark />)).toMatchSnapshot()
-  })
-})
