+++ packages/fyndiq-icons/package.json
@@ -1,23 +1,23 @@
 {
-  "name": "fyndiq-icon-star",
-  "version": "2.0.1",
-  "author": "Thibaut REMY <thibaut.remy@fyndiq.com>",
+  "name": "fyndiq-icons",
+  "version": "2.0.0",
+  "author": "Thibaut Remy <thibaut.remy@fyndiq.com>",
   "main": "lib/index.js",
   "license": "MIT",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/fyndiq/fyndiq-ui.git"
   },
-  "homepage": "https://github.com/fyndiq/fyndiq-ui",
+  "homepage": "http://developers.fyndiq.com/fyndiq-ui/?selectedKind=Icons&selectedStory=default",
   "dependencies": {
-    "fyndiq-styles-colors": "^2.0.0"
+    "fyndiq-styles-colors": "*"
   },
   "peerDependencies": {
     "css-loader": "^0.28.7",
     "postcss-cssnext": "^3.0.2",
     "postcss-import": "^11.0.0",
     "postcss-loader": "^2.0.6",
-    "prop-types": "^15.5.10",
-    "react": "^15.6.1 || ^16.0.0"
+    "prop-types": "^15.0.0",
+    "react": "^15.0.0 || ^16.0.0"
   }
 }
