+++ packages/fyndiq-icons/src/arrow.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import styles from '../styles.css'
+import styles from '../arrow.css'
 
 const Arrow = ({ orientation, className }) => {
   let transform
