+++ packages/fyndiq-icons/src/checkmark.js
