+++ packages/fyndiq-icons/src/index.js
@@ -1,0 +1,9 @@
+import Arrow from './arrow'
+import Star from './star'
+import Checkmark from './checkmark'
+
+export {
+  Arrow,
+  Star,
+  Checkmark,
+}
