+++ packages/fyndiq-icons/star.css
